+++ dto/Request/EggDto.cs
@@ -8,14 +8,14 @@ namespace sipetok_api.dto.Request
         public DateTime production_date { get; set; }
 
         [Required(ErrorMessage = "Stock is required!")]
-        public int stock { get; set; }
+        public double stock { get; set; }
         public int tenant_id { get; set; }
         [Required(ErrorMessage = "Category is required!")]
         public int category_id { get; set; }
 
         public EggDto() { }
 
-        public EggDto(DateTime production_date, int stock, int tenant_id, int category_id)
+        public EggDto(DateTime production_date, double stock, int tenant_id, int category_id)
         {
             this.production_date = production_date;
             this.stock = stock;
